requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
PyMySQL
sqlparse
djangorestframework-simplejwt
python-dotenv
cryptography #Esto es usa la ultima versión de los paquetes
gunicorn


#pip install -r requirements.txt de esta manera se corre el requirement
#source venv/bin/activate   # Linux/macOS
#.\venv\Scripts\activate      windows el source es para trabajar dentro del env